requirements.txt
FastAPI
uvicorn
requests
openai
python-dotenv
